基础接口完整版

+++ controllers/video.go
@@ -4,7 +4,6 @@ import (
 	"ByteRhythm/models"
 	"ByteRhythm/object"
 	"ByteRhythm/utils"
-	"fmt"
 	"os"
 	"time"
 )
@@ -24,24 +23,15 @@ func (c *VideoController) Feed() {
 		latestTimeStamp = int(time.Now().Unix())
 	}
 	latestTime := time.Unix(int64(latestTimeStamp), 0)
-	fmt.Println(latestTime)
+	//不需要验证token，不登录也能看视频流
 	token := c.GetString("token")
 	if token != "" {
-		if err := utils.ValidateToken(token); err != nil {
-			c.Data["json"] = map[string]interface{}{
-				"status_code": 1,
-				"status_msg":  "token验证失败",
-				"video_list":  nil,
-				"next_time":   nil,
-			}
-			c.ServeJSON()
-			return
-		} else {
-			c.o.QueryTable(new(models.Video)).Filter("create_time__lte", latestTime).OrderBy("-create_time").Limit(30, 0).All(&videos)
-		}
+		c.o.QueryTable(new(models.Video)).Filter("create_time__lte", latestTime).OrderBy("-create_time").Limit(30, 0).All(&videos)
+	} else {
+		//经过测试，如果不登录，提交的latestTimeStamp为55574-03-08 08:53:51 +0800 CST，这个数据毫无意义还会干扰查询
+		c.o.QueryTable(new(models.Video)).OrderBy("-create_time").Limit(30, 0).All(&videos)
 	}
-	//经过测试，如果不登录，提交的latestTimeStamp为55574-03-08 08:53:51 +0800 CST，这个数据毫无意义还会干扰查询
-	c.o.QueryTable(new(models.Video)).OrderBy("-create_time").Limit(30, 0).All(&videos)
+
 	if len(videos) == 0 {
 		c.Data["json"] = map[string]interface{}{
 			"status_code": 1,
@@ -91,11 +81,7 @@ func (c *VideoController) Publish() {
 	token := c.GetString("token")
 	title := c.GetString("title")
 	if err := utils.ValidateToken(token); err != nil {
-		c.Data["json"] = map[string]interface{}{
-			"status_code": 1,
-			"status_msg":  "token验证失败",
-		}
-		c.ServeJSON()
+		c.PublishFail("token验证失败")
 		return
 	}
 	var user models.User
@@ -103,16 +89,22 @@ func (c *VideoController) Publish() {
 	username, _ := utils.GetUsernameFromToken(token)
 	c.o.QueryTable(new(models.User)).Filter("username", username).One(&user)
 	if url := c.UploadMP4(c.GetFile("data")); url == "" {
-		c.Data["json"] = map[string]interface{}{
-			"status_code": 1,
-			"status_msg":  "发布失败",
-		}
-		c.ServeJSON()
+		c.PublishFail("发布失败")
 		return
 	} else {
 		imgPath := utils.VideoGetNetImgCount(1, url)
+		if imgPath == "" {
+			c.PublishFail("发布失败")
+			return
+
+		}
 		coverUrl := c.UploadJPG(imgPath, url)
+		if coverUrl == "" {
+			c.PublishFail("发布失败")
+			return
+		}
 		os.Remove(imgPath)
+
 		video := models.Video{
 			AuthorId: &user,
 			PlayUrl:  url,
@@ -120,11 +112,7 @@ func (c *VideoController) Publish() {
 			CoverUrl: coverUrl,
 		}
 		if _, err := c.o.Insert(&video); err != nil {
-			c.Data["json"] = map[string]interface{}{
-				"status_code": 1,
-				"status_msg":  "发布失败",
-			}
-			c.ServeJSON()
+			c.PublishFail("发布失败")
 			return
 		} else {
 			c.Data["json"] = map[string]interface{}{
@@ -137,6 +125,14 @@ func (c *VideoController) Publish() {
 
 	}
 
+}
+
+func (c *VideoController) PublishFail(msg string) {
+	c.Data["json"] = map[string]interface{}{
+		"status_code": 1,
+		"status_msg":  msg,
+	}
+	c.ServeJSON()
 }
 
 // List 获取发布视频列表
